+++ src/frontend/cli.rs
@@ -19,6 +19,8 @@ use clarity_repl::repl;
 use clap::Parser;
 use toml;
 
+#[cfg(feature = "telemetry")]
+use super::telemetry::{telemetry_report_event, DeveloperUsageDigest, DeveloperUsageEvent};
 #[derive(Parser, PartialEq, Clone, Debug)]
 #[clap(version = option_env!("CARGO_PKG_VERSION").expect("Unable to detect version"))]
 struct Opts {
@@ -76,6 +78,9 @@ enum Contract {
 struct GenerateProject {
     /// Project's name
     pub name: String,
+    /// Enable developer usage telemetry
+    #[clap(long = "disable-telemetry")]
+    pub disable_telemetry: Option<bool>,
 }
 
 #[derive(Parser, PartialEq, Clone, Debug)]
@@ -198,7 +203,31 @@ struct Check {
 }
 
 pub fn main() {
-    let args: Opts = Opts::parse();
+    let opts: Opts = match Opts::try_parse() {
+        Ok(opts) => opts,
+        Err(_e) => {
+            if _e.kind == clap::ErrorKind::UnknownArgument {
+                match get_manifest_path(None) {
+                    Some(manifest_path) => {
+                        let manifest = ProjectManifest::from_path(&manifest_path);
+                        if manifest.project.telemetry {
+                            #[cfg(feature = "telemetry")]
+                            telemetry_report_event(DeveloperUsageEvent::UnknownCommand(
+                                DeveloperUsageDigest::new(
+                                    &manifest.project.name,
+                                    &manifest.project.authors,
+                                ),
+                                format!("{}", _e),
+                            ));
+                        }
+                    }
+                    None => {}
+                };
+            }
+            println!("{}", _e);
+            process::exit(1);
+        }
+    };
 
     let hints_enabled = if env::var("CLARINET_DISABLE_HINTS") == Ok("1".into()) {
         false
@@ -206,17 +235,56 @@ pub fn main() {
         true
     };
 
-    match args.command {
+    match opts.command {
         Command::New(project_opts) => {
             let current_path = {
                 let current_dir = env::current_dir().expect("Unable to read current directory");
                 current_dir.to_str().unwrap().to_owned()
             };
 
-            let changes = generate::get_changes_for_new_project(current_path, project_opts.name);
+            let telemetry_enabled = if cfg!(feature = "telemetry") {
+                if let Some(disable_telemetry) = project_opts.disable_telemetry {
+                    !disable_telemetry
+                } else {
+                    println!("{}", yellow!("Send usage data to Hiro."));
+                    println!("{}", yellow!("Help Hiro improve its products and services by automatically sending diagnostics and usage data."));
+                    println!("{}", yellow!("Only high level usage information, and no information identifying you or your project are collected."));
+                    // todo(ludo): once we have a privacy policy available, add a link
+                    // println!("{}", yellow!("Visit http://hiro.so/clarinet-privacy for details."));
+                    println!("{}", yellow!("Enable [Y/n]?"));
+                    let mut buffer = String::new();
+                    std::io::stdin().read_line(&mut buffer).unwrap();
+                    buffer != "n\n"
+                }
+            } else {
+                false
+            };
+            if telemetry_enabled {
+                println!(
+                    "{}",
+                    yellow!("Telemetry enabled. Thanks for helping to improve clarinet!")
+                );
+            } else {
+                println!(
+                    "{}",
+                    yellow!(
+                        "Telemetry disabled. Clarinet will not collect any data on this project."
+                    )
+                );
+            }
+            let project_id = project_opts.name.clone();
+            let changes =
+                generate::get_changes_for_new_project(current_path, project_id, telemetry_enabled);
             execute_changes(changes);
             if hints_enabled {
                 display_post_check_hint();
+            }
+            if telemetry_enabled {
+                #[cfg(feature = "telemetry")]
+                telemetry_report_event(DeveloperUsageEvent::NewProject(DeveloperUsageDigest::new(
+                    &project_opts.name,
+                    &vec![],
+                )));
             }
         }
         Command::Contract(subcommand) => match subcommand {
@@ -271,12 +339,12 @@ pub fn main() {
                     .max_blocking_threads(32)
                     .build()
                     .unwrap();
-
+                let mut retrieved = BTreeSet::new();
                 let res = rt.block_on(session.resolve_link(&repl::settings::InitialLink {
                     contract_id: fork_contract.contract_id.clone(),
                     stacks_node_addr: None,
                     cache: None,
-                }));
+                }, &mut retrieved));
                 let contracts = res.unwrap();
                 let mut changes = vec![];
                 for (contract_id, code, deps) in contracts.into_iter() {
@@ -312,20 +380,31 @@ pub fn main() {
         Command::Poke(cmd) | Command::Console(cmd) => {
             let manifest_path = get_manifest_path_or_exit(cmd.manifest_path);
             let start_repl = true;
-            load_session(&manifest_path, start_repl, &Network::Devnet)
-                .expect("Unable to start REPL");
+            let (_, _, project_manifest, _) =
+                load_session(&manifest_path, start_repl, &Network::Devnet)
+                    .expect("Unable to start REPL");
             if hints_enabled {
                 display_post_poke_hint();
+            }
+            if project_manifest.project.telemetry {
+                #[cfg(feature = "telemetry")]
+                telemetry_report_event(DeveloperUsageEvent::PokeExecuted(
+                    DeveloperUsageDigest::new(
+                        &project_manifest.project.name,
+                        &project_manifest.project.authors,
+                    ),
+                ));
             }
         }
         Command::Check(cmd) => {
             let manifest_path = get_manifest_path_or_exit(cmd.manifest_path);
             let start_repl = false;
-            match load_session(&manifest_path, start_repl, &Network::Devnet) {
+            let project_manifest = match load_session(&manifest_path, start_repl, &Network::Devnet) {
                 Err(e) => {
                     println!("{}", e);
+                    return;
                 }
-                Ok((session, _, output)) => {
+                Ok((session, _, manifest, output)) => {
                     if let Some(message) = output {
                         println!("{}", message);
                     }
@@ -334,29 +413,39 @@ pub fn main() {
                         green!("✔"),
                         session.settings.initial_contracts.len()
                     );
+                    manifest
                 }
-            }
+            };
             if hints_enabled {
                 display_post_check_hint();
+            }
+            if project_manifest.project.telemetry {
+                #[cfg(feature = "telemetry")]
+                telemetry_report_event(DeveloperUsageEvent::CheckExecuted(
+                    DeveloperUsageDigest::new(
+                        &project_manifest.project.name,
+                        &project_manifest.project.authors,
+                    ),
+                ));
             }
         }
         Command::Test(cmd) => {
             let manifest_path = get_manifest_path_or_exit(cmd.manifest_path);
             let start_repl = false;
             let res = load_session(&manifest_path, start_repl, &Network::Devnet);
-            let session = match res {
-                Ok((session, _, output)) => {
+            let (session, project_manifest) = match res {
+                Ok((session, _, manifest, output)) => {
                     if let Some(message) = output {
                         println!("{}", message);
                     }
-                    session
+                    (session, manifest)
                 }
                 Err(e) => {
                     println!("{}", e);
                     return;
                 }
             };
-            run_scripts(
+            let (success, _count) = match run_scripts(
                 cmd.files,
                 cmd.coverage,
                 cmd.costs_report,
@@ -365,9 +454,26 @@ pub fn main() {
                 false,
                 manifest_path,
                 Some(session),
-            );
+            ) {
+                Ok(count) => (true, count),
+                Err((_, count)) => (false, count),
+            };
             if hints_enabled {
                 display_tests_pro_tips_hint();
+            }
+            if project_manifest.project.telemetry {
+                #[cfg(feature = "telemetry")]
+                telemetry_report_event(DeveloperUsageEvent::TestSuiteExecuted(
+                    DeveloperUsageDigest::new(
+                        &project_manifest.project.name,
+                        &project_manifest.project.authors,
+                    ),
+                    success,
+                    _count,
+                ));
+            }
+            if !success {
+                process::exit(1)
             }
         }
         Command::Run(cmd) => {
@@ -375,7 +481,7 @@ pub fn main() {
             let start_repl = false;
             let res = load_session(&manifest_path, start_repl, &Network::Devnet);
             let session = match res {
-                Ok((session, _, output)) => {
+                Ok((session, _, _, output)) => {
                     if let Some(message) = output {
                         println!("{}", message);
                     }
@@ -386,7 +492,7 @@ pub fn main() {
                     return;
                 }
             };
-            run_scripts(
+            let _ = run_scripts(
                 vec![cmd.script],
                 false,
                 false,
@@ -409,14 +515,39 @@ pub fn main() {
             } else {
                 panic!("Target deployment must be specified with --devnet, --testnet or --mainnet")
             };
-            match publish_all_contracts(&manifest_path, network, true, 30) {
-                Ok(results) => println!("{}", results.join("\n")),
-                Err(results) => println!("{}", results.join("\n")),
+            let project_manifest = match publish_all_contracts(&manifest_path, network, true, 30) {
+                Ok((results, project_manifest)) => {
+                    println!("{}", results.join("\n"));
+                    project_manifest
+                }
+                Err(results) => {
+                    println!("{}", results.join("\n"));
+                    return;
+                }
             };
+            if project_manifest.project.telemetry {
+                #[cfg(feature = "telemetry")]
+                telemetry_report_event(DeveloperUsageEvent::ContractPublished(
+                    DeveloperUsageDigest::new(
+                        &project_manifest.project.name,
+                        &project_manifest.project.authors,
+                    ),
+                    network,
+                ));
+            }
         }
         Command::Integrate(cmd) => {
             let manifest_path = get_manifest_path_or_exit(cmd.manifest_path);
             let devnet = DevnetOrchestrator::new(manifest_path, None);
+            if devnet.manifest.project.telemetry {
+                #[cfg(feature = "telemetry")]
+                telemetry_report_event(DeveloperUsageEvent::DevnetExecuted(
+                    DeveloperUsageDigest::new(
+                        &devnet.manifest.project.name,
+                        &devnet.manifest.project.authors,
+                    ),
+                ));
+            }
             let _ = integrate::run_devnet(devnet, None, !cmd.no_dashboard);
             if hints_enabled {
                 display_deploy_hint();
@@ -426,29 +557,37 @@ pub fn main() {
     };
 }
 
-fn get_manifest_path_or_exit(path: Option<String>) -> PathBuf {
-    println!("");
+fn get_manifest_path(path: Option<String>) -> Option<PathBuf> {
     if let Some(path) = path {
         let manifest_path = PathBuf::from(path);
         if !manifest_path.exists() {
-            println!("Could not find Clarinet.toml");
-            process::exit(1);
+            return None;
         }
-        manifest_path
+        Some(manifest_path)
     } else {
         let mut current_dir = env::current_dir().unwrap();
         loop {
             current_dir.push("Clarinet.toml");
 
             if current_dir.exists() {
-                break current_dir;
+                return Some(current_dir);
             }
             current_dir.pop();
 
             if !current_dir.pop() {
-                println!("Could not find Clarinet.toml");
-                process::exit(1);
+                return None;
             }
+        }
+    }
+}
+
+fn get_manifest_path_or_exit(path: Option<String>) -> PathBuf {
+    println!("");
+    match get_manifest_path(path) {
+        Some(manifest_path) => manifest_path,
+        None => {
+            println!("Could not find Clarinet.toml");
+            process::exit(1);
         }
     }
 }
@@ -554,7 +693,7 @@ fn display_post_check_hint() {
         "{}",
         yellow!("    Run all run tests in the ./tests folder.\n")
     );
-    println!("{}", yellow!("Find more information on testing with Clarinet here: https://docs.hiro.so/docs/smart-contracts/clarinet#testing-with-the-test-harness"));
+    println!("{}", yellow!("Find more information on testing with Clarinet here: https://docs.hiro.so/smart-contracts/clarinet#testing-with-the-test-harness"));
     display_hint_footer();
 }
 
@@ -577,7 +716,7 @@ fn display_post_poke_hint() {
         yellow!("    Check contract syntax for all files in ./contracts.\n")
     );
 
-    println!("{}", yellow!("Find more information on writing contracts with Clarinet here: https://docs.hiro.so/docs/smart-contracts/clarinet#developing-a-clarity-smart-contract"));
+    println!("{}", yellow!("Find more information on writing contracts with Clarinet here: https://docs.hiro.so/smart-contracts/clarinet#developing-a-clarity-smart-contract"));
     display_hint_footer();
 }
 
@@ -615,7 +754,7 @@ fn display_tests_pro_tips_hint() {
         yellow!("    Deploy all contracts to a local dockerized blockchain setup (Devnet).\n")
     );
 
-    println!("{}", yellow!("Find more information on testing with Clarinet here: https://docs.hiro.so/docs/smart-contracts/clarinet#testing-with-clarinet"));
+    println!("{}", yellow!("Find more information on testing with Clarinet here: https://docs.hiro.so/smart-contracts/clarinet#testing-with-clarinet"));
     display_hint_footer();
 }
 
@@ -645,7 +784,7 @@ fn display_deploy_hint() {
     );
     println!(
         "{}",
-        yellow!("Find more information on the DevNet here: https://docs.hiro.so/docs/smart-contracts/devnet")
+        yellow!("Find more information on the DevNet here: https://docs.hiro.so/smart-contracts/devnet/")
     );
     display_hint_footer();
 }