+++ src/frontend/mod.rs
@@ -1,1 +1,3 @@
 pub mod cli;
+#[cfg(feature = "telemetry")]
+mod telemetry;