+++ src/integrate/events_observer.rs
@@ -72,7 +72,7 @@ impl EventObserverConfig {
     pub fn new(devnet_config: DevnetConfig, manifest_path: PathBuf) -> Self {
         info!("Checking contracts...");
         let (session, config) = match load_session(&manifest_path, false, &Network::Devnet) {
-            Ok((session, config, _)) => (session, config),
+            Ok((session, config, _, _)) => (session, config),
             Err(e) => {
                 println!("{}", e);
                 std::process::exit(1);
@@ -113,7 +113,6 @@ impl EventObserverConfig {
 
 pub enum EventsObserverCommand {
     Terminate(bool), // Restart
-    UpdatePoxInfo,
     PublishInitialContracts,
     PublishPoxStackingOrders(BlockIdentifier),
 }
@@ -176,6 +175,7 @@ pub async fn start_events_observer(
             .mount(
                 "/",
                 routes![
+                    handle_ping,
                     handle_new_burn_block,
                     handle_new_block,
                     handle_new_microblocks,
@@ -205,7 +205,7 @@ pub async fn start_events_observer(
                     .expect("Unable to terminate event observer");
 
                 let session = match load_session(&manifest_path, false, &Network::Devnet) {
-                    Ok((session, _, _)) => session,
+                    Ok((session, _, _, _)) => session,
                     Err(e) => {
                         devnet_event_tx
                             .send(DevnetEvent::error(format!("Contracts invalid: {}", e)))
@@ -230,7 +230,6 @@ pub async fn start_events_observer(
                     init_status.deployer_nonce = 0;
                 }
             }
-            Ok(EventsObserverCommand::UpdatePoxInfo) => {}
             Ok(EventsObserverCommand::PublishInitialContracts) => {
                 if !contracts_deployed {
                     contracts_deployed = true;