+++ CHANGELOG.md
@@ -1,3 +1,89 @@
+# [0.24.0](https://github.com/hirosystems/clarinet/compare/v0.23.1...v0.24.0) (2022-01-19)
+
+
+### Bug Fixes
+
+* fixed broken links ([b0f073a](https://github.com/hirosystems/clarinet/commit/b0f073ac634f480c86b0b788c51057605b202c64))
+* generate proper strings from contract results ([6b189c6](https://github.com/hirosystems/clarinet/commit/6b189c61b36f1f1ca755b86225916373565411bb)), closes [#213](https://github.com/hirosystems/clarinet/issues/213)
+* only code-sign on release ([e10f3d1](https://github.com/hirosystems/clarinet/commit/e10f3d12eeae10e0fe809df4b2bdfbe97d33a26a))
+* resolve bug with windows build in CI ([16ccd00](https://github.com/hirosystems/clarinet/commit/16ccd00110c6abde8f4e27b74ddd93f935028e36))
+
+
+### Features
+
+* added ignore and only options to clarinet test ([a57cc23](https://github.com/hirosystems/clarinet/commit/a57cc2373cfe25604290c59143469df649337657))
+
+## [0.23.1](https://github.com/hirosystems/clarinet/compare/v0.23.0...v0.23.1) (2022-01-13)
+
+
+### Bug Fixes
+
+* issue with chunked information in Mempool section ([1dd1e22](https://github.com/hirosystems/clarinet/commit/1dd1e22e1f14f6d23a8cdfe364bebb88489b25cb))
+
+# [0.23.0](https://github.com/hirosystems/clarinet/compare/v0.22.0...v0.23.0) (2022-01-07)
+
+
+### Bug Fixes
+
+* display warnings and notes ([e0c4e1c](https://github.com/hirosystems/clarinet/commit/e0c4e1c0a1bd41a1814225a2565607523cfb25be))
+* off by one spans ([c724911](https://github.com/hirosystems/clarinet/commit/c724911e1f252c123d382947576014fa648638ee))
+
+
+### Features
+
+* display warning as warning in popups ([0ec64cf](https://github.com/hirosystems/clarinet/commit/0ec64cf4ea53ac169c7f798edca69be24d0c97f9))
+
+# [0.22.0](https://github.com/hirosystems/clarinet/compare/v0.21.2...v0.22.0) (2022-01-05)
+
+
+### Bug Fixes
+
+* lsp adjustment ([67233f2](https://github.com/hirosystems/clarinet/commit/67233f29fb01a4f3faa192d1a0d62866471605b1))
+
+
+### Features
+
+* fail gracefully on Clarinet.toml errors ([3023320](https://github.com/hirosystems/clarinet/commit/3023320cfe919f925fbb0710d42811d12100e430))
+
+## [0.21.2](https://github.com/hirosystems/clarinet/compare/v0.21.1...v0.21.2) (2021-12-23)
+
+
+### Bug Fixes
+
+* rollback to clarity-repl v0.18.0 ([c5db67b](https://github.com/hirosystems/clarinet/commit/c5db67be42915492f3a619156611d19133d5fb82))
+
+## [0.21.1](https://github.com/hirosystems/clarinet/compare/v0.21.0...v0.21.1) (2021-12-23)
+
+
+### Bug Fixes
+
+* show errors when parsing opts ([2dad960](https://github.com/hirosystems/clarinet/commit/2dad96064c95b763a1df13b2ea56421acc527a27)), closes [#188](https://github.com/hirosystems/clarinet/issues/188)
+
+# [0.21.0](https://github.com/hirosystems/clarinet/compare/v0.20.0...v0.21.0) (2021-12-22)
+
+
+### Bug Fixes
+
+* build error ([6399169](https://github.com/hirosystems/clarinet/commit/63991693e3f9ce33c7ee010a5419190d0ed2c7cf))
+* caveman debug vestige ([a89c631](https://github.com/hirosystems/clarinet/commit/a89c631e3cd283094bd81d51c41379acb9bf28e3))
+* clarity integration ([259fb4a](https://github.com/hirosystems/clarinet/commit/259fb4a0669d0c0e58b4192c1a21e187cff8572a))
+* comments and doc ([ea269ae](https://github.com/hirosystems/clarinet/commit/ea269aefc980b6aea663ff1115cec4321c13179c))
+* disable_telemetry flag ([8dcb0ec](https://github.com/hirosystems/clarinet/commit/8dcb0ec805a66570d8e10d1bb1e05e36e7cae43b))
+* doc copy pasta ([8927144](https://github.com/hirosystems/clarinet/commit/8927144e7240049beadbfbc19c2671ba299d523f))
+* enable telemetry in Dockerfile ([8b43980](https://github.com/hirosystems/clarinet/commit/8b43980a18ac9d996449b7873384236ae3be5c40))
+* iterate on integration ([3a65165](https://github.com/hirosystems/clarinet/commit/3a651655b2aa1a39aafac6a38a83a082a612b48c))
+* make tower-lsp optional ([1f83b21](https://github.com/hirosystems/clarinet/commit/1f83b2146205ed3ad8479c865ef8c2e0deb523bf))
+* remove reference to Blockstack ([350be75](https://github.com/hirosystems/clarinet/commit/350be75a8368bf27dea36fa14275e24d5f3feca4))
+* stacks-devnet-js build ([fe74897](https://github.com/hirosystems/clarinet/commit/fe74897b03b2b5a464e8f6a874f0a2f3a6311373))
+
+
+### Features
+
+* add `analysis` field into project config ([ce61daf](https://github.com/hirosystems/clarinet/commit/ce61dafb92048e8268cc0082250d838b9289163f))
+* add key ([0631e72](https://github.com/hirosystems/clarinet/commit/0631e720d4e4260c94a725b4f78fbb0dad5cf242))
+* implement telemetry as a gated feature ([8b04f8b](https://github.com/hirosystems/clarinet/commit/8b04f8ba5f1cb0017f28ab2f050cf68f2af8369b))
+* implement telemetry helpers ([b1cbcc6](https://github.com/hirosystems/clarinet/commit/b1cbcc6d264f5dc7f49fef983ef167b330028e06))
+
 # [0.20.0](https://github.com/hirosystems/clarinet/compare/v0.19.1...v0.20.0) (2021-12-09)